+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/svelte
 
+## 5.0.35
+
+### Patch Changes
+
+- Updated dependencies [7f6650b]
+- Updated dependencies [106ea59]
+  - ai@7.0.35
+
 ## 5.0.34
 
 ### Patch Changes